+++ copulas/__init__.py
@@ -4,11 +4,17 @@
 
 __author__ = 'MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com',
-__version__ = '0.2.1-dev'
+__version__ = '0.2.2-dev'
+
+import importlib
 
 import numpy as np
 
 EPSILON = np.finfo(np.float32).eps
+
+
+class NotFittedError(Exception):
+    pass
 
 
 def random_state(function):
@@ -26,3 +32,21 @@ def random_state(function):
             return result
 
     return wrapper
+
+
+def import_object(object_name):
+    """Import an object from its Fully Qualified Name."""
+    package, name = object_name.rsplit('.', 1)
+    return getattr(importlib.import_module(package), name)
+
+
+def get_qualified_name(_object):
+    """Return the Fully Qualified Name from an instance or class."""
+    module = _object.__module__
+    if hasattr(_object, '__name__'):
+        _class = _object.__name__
+
+    else:
+        _class = _object.__class__.__name__
+
+    return module + '.' + _class