+++ copulas/multivariate/tree.py
@@ -5,8 +5,9 @@ from enum import Enum
 import numpy as np
 import scipy
 
-from copulas import EPSILON
+from copulas import EPSILON, get_qualified_name
 from copulas.bivariate.base import Bivariate
+from copulas.multivariate.base import Multivariate
 
 LOGGER = logging.getLogger(__name__)
 
@@ -17,7 +18,7 @@ class TreeTypes(Enum):
     REGULAR = 2
 
 
-class Tree(object):
+class Tree(Multivariate):
     """Helper class to instantiate a single tree in the vine model."""
 
     tree_type = None
@@ -57,7 +58,8 @@ class Tree(object):
         Args:
             tree_type: `TreeType` or `str` to be compared against  TreeType.
         """
-        pass
+        self.tree_type = tree_type
+        self.fitted = False
 
     def fit(self, index, n_nodes, tau_matrix, previous_tree, edges=None):
         """Fits tree object.
@@ -86,6 +88,8 @@ class Tree(object):
                 self._build_kth_tree()
 
             self.prepare_next_tree()
+
+        self.fitted = True
 
     def _check_contraint(self, edge1, edge2):
         """Check if two edges satisfy vine constraint.
@@ -204,7 +208,7 @@ class Tree(object):
             right_given_left[right_given_left == 0] = EPSILON
             left_given_right[left_given_right == 1] = 1 - EPSILON
             right_given_left[right_given_left == 1] = 1 - EPSILON
-            edge.U = [left_given_right, right_given_left]
+            edge.U = np.array([left_given_right, right_given_left])
 
     def get_likelihood(self, uni_matrix):
         """Compute likelihood of the tree given an U matrix.
@@ -238,19 +242,55 @@ class Tree(object):
         return '\n'.join([template.format(edge.L, edge.R, edge.D, edge.name, edge.theta)
                           for edge in self.edges])
 
-    def to_dict(self):
-        return {
-            'level': self.level - 1,
-            'n_nodes': self.n_nodes,
-            'tau_matrix': self.tau_matrix.tolist(),
-            'previous_tree': self.previous_tree.tolist(),
-            'edges': [edge.to_dict() for edge in self.edges],
-        }
+    def _serialize_previous_tree(self):
+        if self.level == 1:
+            return self.previous_tree.tolist()
+
+        return None
 
     @classmethod
-    def from_dict(cls, tree_dict):
+    def _deserialize_previous_tree(cls, tree_dict, previous):
+        if tree_dict['level'] == 1:
+            return np.array(tree_dict['previous_tree'])
+
+        return previous
+
+    def to_dict(self):
+        fitted = self.fitted
+        result = {
+            'tree_type': self.tree_type,
+            'type': get_qualified_name(self),
+            'fitted': fitted
+        }
+
+        if not fitted:
+            return result
+
+        result.update({
+            'level': self.level,
+            'n_nodes': self.n_nodes,
+            'tau_matrix': self.tau_matrix.tolist(),
+            'previous_tree': self._serialize_previous_tree(),
+            'edges': [edge.to_dict() for edge in self.edges],
+        })
+
+        return result
+
+    @classmethod
+    def from_dict(cls, tree_dict, previous=None):
         """Create a new instance from a dictionary."""
-        return cls(**tree_dict)
+        instance = cls(tree_dict['tree_type'])
+
+        fitted = tree_dict['fitted']
+        instance.fitted = fitted
+        if fitted:
+            instance.level = tree_dict['level']
+            instance.n_nodes = tree_dict['n_nodes']
+            instance.tau_matrix = np.array(tree_dict['tau_matrix'])
+            instance.previous_tree = cls._deserialize_previous_tree(tree_dict, previous)
+            instance.edges = [Edge.from_dict(edge) for edge in tree_dict['edges']]
+
+        return instance
 
     @classmethod
     def load(cls, tree_path):
@@ -279,7 +319,7 @@ class CenterTree(Tree):
             ind = int(tau_sorted[itr, 0])
             name, theta = Bivariate.select_copula(self.u_matrix[:, (0, ind)])
 
-            new_edge = Edge(0, ind, name, theta)
+            new_edge = Edge(itr, 0, ind, name, theta)
             new_edge.tau = self.tau_matrix[0, ind]
             self.edges.append(new_edge)
 
@@ -292,7 +332,7 @@ class CenterTree(Tree):
         for itr in range(self.n_nodes - 1):
             right = int(aux_sorted[itr, 0])
             left_parent, right_parent = Edge.sort_edge([edges[anchor], edges[right]])
-            new_edge = Edge.get_child_edge(left_parent, right_parent)
+            new_edge = Edge.get_child_edge(itr, left_parent, right_parent)
             new_edge.tau = aux_sorted[itr, 1]
             self.edges.append(new_edge)
 
@@ -343,7 +383,7 @@ class DirectTree(Tree):
             name, theta = Bivariate.select_copula(self.u_matrix[:, (T1[k], T1[k + 1])])
 
             left, right = sorted([T1[k], T1[k + 1]])
-            new_edge = Edge(left, right, name, theta)
+            new_edge = Edge(k, left, right, name, theta)
             new_edge.tau = tau_T1[k]
             self.edges.append(new_edge)
 
@@ -351,7 +391,7 @@ class DirectTree(Tree):
         edges = self.previous_tree.edges
         for k in range(self.n_nodes - 1):
             left_parent, right_parent = Edge.sort_edge([edges[k], edges[k + 1]])
-            new_edge = Edge.get_child_edge(left_parent, right_parent)
+            new_edge = Edge.get_child_edge(k, left_parent, right_parent)
             new_edge.tau = self.tau_matrix[k, k + 1]
             self.edges.append(new_edge)
 
@@ -379,7 +419,7 @@ class RegularTree(Tree):
             name, theta = Bivariate.select_copula(self.u_matrix[:, (edge[0], edge[1])])
 
             left, right = sorted([edge[0], edge[1]])
-            new_edge = Edge(left, right, name, theta)
+            new_edge = Edge(len(X) - 1, left, right, name, theta)
             new_edge.tau = self.tau_matrix[edge[0], edge[1]]
             self.edges.append(new_edge)
             X.add(edge[1])
@@ -407,7 +447,7 @@ class RegularTree(Tree):
             pairs = sorted(adj_set, key=lambda e: neg_tau[e[0]][e[1]])[0]
             left_parent, right_parent = Edge.sort_edge([edges[pairs[0]], edges[pairs[1]]])
 
-            new_edge = Edge.get_child_edge(left_parent, right_parent)
+            new_edge = Edge.get_child_edge(len(visited) - 1, left_parent, right_parent)
             new_edge.tau = self.tau_matrix[pairs[0], pairs[1]]
             self.edges.append(new_edge)
 
@@ -416,7 +456,7 @@ class RegularTree(Tree):
 
 
 class Edge(object):
-    def __init__(self, left, right, copula_name, copula_theta):
+    def __init__(self, index, left, right, copula_name, copula_theta):
         """Initialize an Edge object.
 
         Args:
@@ -426,6 +466,7 @@ class Edge(object):
             :param copula_theta: parameters of the fitted copula class
 
         """
+        self.index = index
         self.L = left
         self.R = right
         self.D = set()  # dependence_set
@@ -497,13 +538,13 @@ class Edge(object):
         return left_u, right_u
 
     @classmethod
-    def get_child_edge(cls, left_parent, right_parent):
+    def get_child_edge(cls, index, left_parent, right_parent):
         """Construct a child edge from two parent edges."""
         [ed1, ed2, depend_set] = cls._identify_eds_ing(left_parent, right_parent)
         left_u, right_u = cls.get_conditional_uni(left_parent, right_parent)
         X = np.array([[x, y] for x, y in zip(left_u, right_u)])
         name, theta = Bivariate.select_copula(X)
-        new_edge = Edge(ed1, ed2, name, theta)
+        new_edge = Edge(index, ed1, ed2, name, theta)
         new_edge.D = depend_set
         new_edge.parents = [left_parent, right_parent]
         return new_edge
@@ -537,28 +578,32 @@ class Edge(object):
         if self.parents:
             parents = [parent.to_dict() for parent in self.parents]
 
-        neighbors = None
-        if self.neighbors:
-            neighbors = [neighbor.to_dict() for neighbor in self.neighbors]
+        U = None
+        if self.U is not None:
+            U = self.U.tolist()
 
         return {
+            'index': self.index,
             'L': self.L,
             'R': self.R,
             'D': self.D,
             'parents': parents,
-            'neighbors': neighbors,
+            'neighbors': self.neighbors,
             'name': self.name,
             'theta': self.theta,
             'tau': self.tau,
-            'U': self.U,
+            'U': U,
             'likelihood': self.likelihood
         }
 
     @classmethod
     def from_dict(cls, edge_dict):
-        instance = cls(edge_dict['L'], edge_dict['R'], edge_dict['name'], edge_dict['theta'])
+        instance = cls(
+            edge_dict['index'], edge_dict['L'], edge_dict['R'],
+            edge_dict['name'], edge_dict['theta']
+        )
+        instance.U = np.array(edge_dict['U'])
         parents = edge_dict['parents']
-        neighbors = edge_dict['neighbors']
 
         if parents:
             instance.parents = []
@@ -566,13 +611,7 @@ class Edge(object):
                 edge = Edge.from_dict(parent)
                 instance.parents.append(edge)
 
-        if neighbors:
-            instance.neighbors = []
-            for neighbor in neighbors:
-                edge = Edge.from_dict(neighbor)
-                instance.neighbors.append(edge)
-
-        regular_attributes = ['D', 'tau', 'U', 'likelihood']
+        regular_attributes = ['D', 'tau', 'likelihood', 'neighbors']
         for key in regular_attributes:
             setattr(instance, key, edge_dict[key])