+++ copulas/univariate/gaussian.py
@@ -13,6 +13,8 @@ LOGGER = logging.getLogger(__name__)
 class GaussianUnivariate(Univariate):
     """Gaussian univariate model."""
 
+    fitted = False
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.name = None
@@ -48,6 +50,7 @@ class GaussianUnivariate(Univariate):
 
         self.mean = np.mean(X)
         self.std = np.std(X) or 0.001
+        self.fitted = True
 
     def probability_density(self, X):
         """Compute probability density.
@@ -58,6 +61,7 @@ class GaussianUnivariate(Univariate):
         Returns:
             np.ndarray
         """
+        self.check_fit()
         return norm.pdf(X, loc=self.mean, scale=self.std)
 
     def cumulative_distribution(self, X):
@@ -69,7 +73,7 @@ class GaussianUnivariate(Univariate):
         Returns:
             np.ndarray: Cumulative density for X.
         """
-
+        self.check_fit()
         return norm.cdf(X, loc=self.mean, scale=self.std)
 
     def percent_point(self, U):
@@ -81,6 +85,7 @@ class GaussianUnivariate(Univariate):
         Returns:
             `np.ndarray`: Estimated values in original space.
         """
+        self.check_fit()
         return norm.ppf(U, loc=self.mean, scale=self.std)
 
     @random_state
@@ -93,18 +98,23 @@ class GaussianUnivariate(Univariate):
         Returns:
             np.ndarray: Generated samples
         """
+        self.check_fit()
         return np.random.normal(self.mean, self.std, num_samples)
 
-    def to_dict(self):
+    def _fit_params(self):
         return {
             'mean': self.mean,
-            'std': self.std
+            'std': self.std,
         }
 
     @classmethod
     def from_dict(cls, copula_dict):
         """Set attributes with provided values."""
         instance = cls()
-        instance.mean = copula_dict['mean']
-        instance.std = copula_dict['std']
+        instance.fitted = copula_dict['fitted']
+
+        if instance.fitted:
+            instance.mean = copula_dict['mean']
+            instance.std = copula_dict['std']
+
         return instance