+++ tests/copulas/bivariate/test_base.py
@@ -83,8 +83,9 @@ class TestBivariate(TestCase):
         # Check
         assert result == expected_result
 
+    @mock.patch("builtins.open")
     @mock.patch('copulas.bivariate.base.json.dump')
-    def test_save(self, json_mock):
+    def test_save(self, json_mock, open_mock):
         """Save stores the internal dictionary as a json in a file."""
         # Setup
         instance = Bivariate('frank')
@@ -100,12 +101,13 @@ class TestBivariate(TestCase):
         instance.save('test.json')
 
         # Check
+        assert open_mock.called_once_with('test.json', 'w')
         assert json_mock.called
         compare_nested_dicts(json_mock.call_args[0][0], expected_content)
 
-    @mock.patch('builtins.open', new_callable=mock.mock_open)
+    @mock.patch('builtins.open')
     @mock.patch('copulas.bivariate.base.json.load')
-    def test_load_from_file(self, json_mock, file_mock):
+    def test_load_from_file(self, json_mock, open_mock):
         """Load can recreate an instance from a saved file."""
         # Setup
         json_mock.return_value = {
@@ -118,6 +120,7 @@ class TestBivariate(TestCase):
         instance = Bivariate.load('somefile.json')
 
         # Check
+        assert open_mock.called_once_with('test.json', 'r')
         instance.copula_type == CopulaTypes.FRANK
         instance.tau == -0.33333333333333337
         instance.theta == -3.305771759329249
@@ -139,3 +142,22 @@ class TestBivariate(TestCase):
 
         # Check
         assert name == expected
+
+    @mock.patch('copulas.bivariate.clayton.Clayton.partial_derivative')
+    def test_partial_derivative_scalar(self, derivative_mock):
+        """partial_derivative_scalar calls partial_derivative with its arguments in an array."""
+        # Setup
+        instance = Bivariate(CopulaTypes.CLAYTON)
+        instance.fit(self.X)
+
+        # Run
+        result = instance.partial_derivative_scalar(0.5, 0.1)
+
+        # Check
+        assert result == derivative_mock.return_value
+
+        expected_args = ((np.array([[0.5, 0.1]]), 0), {})
+        assert len(expected_args) == len(derivative_mock.call_args)
+        assert (derivative_mock.call_args[0][0] == expected_args[0][0]).all()
+        assert derivative_mock.call_args[0][1] == expected_args[0][1]
+        assert derivative_mock.call_args[1] == expected_args[1]