+++ tests/copulas/multivariate/test_vine.py
@@ -1,9 +1,13 @@
-from unittest import TestCase, expectedFailure
+from unittest import TestCase
 
 import numpy as np
 import pandas as pd
 
+from copulas.multivariate.base import Multivariate
+from copulas.multivariate.tree import Tree, TreeTypes
 from copulas.multivariate.vine import VineCopula
+from copulas.univariate import KDEUnivariate
+from tests import compare_nested_dicts
 
 
 class TestVine(TestCase):
@@ -35,13 +39,13 @@ class TestVine(TestCase):
             ])
         })
 
-        self.rvine = VineCopula('regular')
+        self.rvine = VineCopula(TreeTypes.REGULAR)
         self.rvine.fit(data)
 
-        self.cvine = VineCopula('center')
+        self.cvine = VineCopula(TreeTypes.CENTER)
         self.cvine.fit(data)
 
-        self.dvine = VineCopula('direct')
+        self.dvine = VineCopula(TreeTypes.DIRECT)
         self.dvine.fit(data)
 
     def test_get_likelihood(self):
@@ -60,12 +64,162 @@ class TestVine(TestCase):
         expected = -0.27565584158521045
         assert abs(dvalue - expected) < 10E-3
 
-    @expectedFailure
-    def test_sample(self):
-        sample_r = self.rvine.sample()
-        sample_c = self.cvine.sample()
-        sample_d = self.cvine.sample()
+    def test_to_dict(self):
+        """ """
+        # Setup
+        instance = VineCopula('regular')
+        instance.fitted = True
+        instance.n_sample = 100
+        instance.n_var = 10
+        instance.depth = 3
+        instance.truncated = 3
+        tree = Tree('regular')
+        instance.trees = [tree]
+        uni = KDEUnivariate()
+        instance.unis = [uni]
 
-        assert len(sample_r) == 4
-        assert len(sample_c) == 4
-        assert len(sample_d) == 4
+        tau_mat = np.array([
+            [0, 1],
+            [1, 0]
+        ])
+        instance.tau_mat = tau_mat
+
+        u_matrix = np.array([
+            [0, 1],
+            [1, 0]
+        ])
+        instance.u_matrix = u_matrix
+
+        expected_result = {
+            'type': 'copulas.multivariate.vine.VineCopula',
+            'fitted': True,
+            'vine_type': 'regular',
+            'n_sample': 100,
+            'n_var': 10,
+            'depth': 3,
+            'truncated': 3,
+            'trees': [
+                {
+                    'type': 'copulas.multivariate.tree.RegularTree',
+                    'tree_type': 'regular',
+                    'fitted': False
+                }
+            ],
+            'tau_mat': [
+                [0, 1],
+                [1, 0]
+            ],
+            'u_matrix': [
+                [0, 1],
+                [1, 0]
+            ],
+            'unis': [
+                {
+                    'type': 'copulas.univariate.kde.KDEUnivariate',
+                    'fitted': False
+                }
+            ]
+        }
+
+        # Run
+        result = instance.to_dict()
+
+        # Check
+        assert result == expected_result
+
+    def test_from_dict(self):
+        # Setup
+        vine_dict = {
+            'type': 'copulas.multivariate.vine.VineCopula',
+            'vine_type': 'regular',
+            'fitted': True,
+            'n_sample': 100,
+            'n_var': 10,
+            'depth': 3,
+            'truncated': 3,
+            'trees': [
+                {
+                    'type': 'copulas.multivariate.tree.RegularTree',
+                    'tree_type': 'regular',
+                    'fitted': False
+                }
+            ],
+            'tau_mat': [
+                [0, 1],
+                [1, 0]
+            ],
+            'u_matrix': [
+                [0, 1],
+                [1, 0]
+            ],
+            'unis': [
+                {
+                    'type': 'copulas.univariate.kde.KDEUnivariate',
+                    'fitted': False
+                }
+            ]
+        }
+
+        # Run
+        instance = Multivariate.from_dict(vine_dict)
+
+        # Check
+        assert instance.vine_type == 'regular'
+        assert instance.n_sample == 100
+        assert instance.n_var == 10
+        assert instance.depth == 3
+        assert instance.truncated == 3
+        assert len(instance.trees) == 1
+        assert instance.trees[0].to_dict() == Tree('regular').to_dict()
+        assert (instance.tau_mat == np.array([
+            [0, 1],
+            [1, 0]
+        ])).all()
+        assert (instance.u_matrix == np.array([
+            [0, 1],
+            [1, 0]
+        ])).all()
+
+    def test_serialization_unfitted_model(self):
+        # Setup
+        instance = VineCopula('regular')
+
+        # Run
+        result = VineCopula.from_dict(instance.to_dict())
+
+        # Check
+        assert result.to_dict() == instance.to_dict()
+
+    def test_serialization_fit_model(self):
+        # Setup
+        instance = VineCopula('regular')
+        X = pd.DataFrame(data=[
+            [1, 0, 0],
+            [0, 1, 0],
+            [0, 0, 1]
+        ])
+        instance.fit(X)
+
+        # Run
+        result = VineCopula.from_dict(instance.to_dict())
+
+        # Check
+        compare_nested_dicts(result.to_dict(), instance.to_dict())
+
+    def test_sample(self):
+        """After being fit, a vine can sample new data."""
+        # Setup
+        vine = VineCopula(TreeTypes.REGULAR)
+        X = pd.DataFrame([
+            [1, 0, 0, 0],
+            [0, 1, 0, 0],
+            [0, 0, 1, 0],
+            [0, 0, 0, 1]
+        ])
+        vine.fit(X)
+
+        # Run
+        result = vine.sample()
+
+        # Check
+        assert len(result) == vine.n_var