+++ HISTORY.md
@@ -1,5 +1,14 @@
 # History
 
+## 0.2.1 - Vine serialization
+
+* Add serialization to Vine copulas.
+* Add `distribution` as argument for the Gaussian Copula.
+* Improve Bivariate Copulas code structure to remove code duplication.
+* Fix bug in Vine Copulas sampling: 'Edge' object has no attribute 'index'
+* Improve code documentation.
+* Improve code style and linting tools configuration.
+
 ## 0.2.0 - Unified API
 
 * New API for stats methods.